Worked on animelistPage to include cards

+++ weeb_roulette/routers/anime.py
@@ -16,9 +16,9 @@ not_authorized = HTTPException(
     headers={"WWW-Authenticate": "Bearer"},
 )
 
-@router.get("/anime_list")
-def get_anime_lists():
-    data = get_anime_list()
+@router.get("/anime_list/{id}")
+def get_anime_lists(id):
+    data = get_anime_list(id)
     return data
 
 @router.get("/anime_detail/{id}")
@@ -61,7 +61,7 @@ def add_anime_to_queue(
         response.status_code = 404
     return record
 
-#remove an anime to user's queue
+#remove an anime from user's queue
 @router.put("/remove_anime_from_queue/{id}", response_model = AnimeQueue)
 def remove_anime_from_queue(
     id: str,
@@ -87,30 +87,30 @@ def update_queue_name(
         response.status_code = 404
     return record
 
-#create new anime queue
+#create new anime queue attached to profile
 @router.post("/anime_queues", response_model=AnimeQueue)
 async def create_anime_queue(
     profile_id: str,
     animequeue: AnimeQueueIn,
     repo: AnimeQueueQueries = Depends(),
-    #  account_data: dict = Depends(authenticator.get_current_account_data),
+    account_data: dict = Depends(authenticator.get_current_account_data),
 ):
-    # account = AccountOut(**account_data)
-    # if "user" not in account.roles:
-    #      raise not_authorized
+    account = AccountOut(**account_data)
+    if "user" not in account.roles:
+         raise not_authorized
     animequeue_request = AnimeQueueIn(profile_id = profile_id, name = animequeue.name)
     animequeue_request = repo.create_queue(animequeue_request)
     return animequeue_request
 
-#delete anime queue
+#delete anime queue attached to profile
 @router.delete("/anime_queues/{id}", response_model=bool)
 async def delete_user_queue(
     id: str,
     repo: AnimeQueueQueries = Depends(),
-    # account_data: dict = Depends(authenticator.get_current_account_data),
+    account_data: dict = Depends(authenticator.get_current_account_data),
 ):
-    # account = AccountOut(**account_data)
-    # if "user" not in account.roles:
-    #     raise not_authorized
+    account = AccountOut(**account_data)
+    if "user" not in account.roles:
+        raise not_authorized
     repo.delete_queue(id=id)
     return True
